% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Genome_Intervals.R
\name{create.genome.intervals}
\alias{create.genome.intervals}
\title{Create Genome Intervals}
\usage{
create.genome.intervals(chromosome.sizes, intervals = TRUE, size = 10)
}
\arguments{
\item{chromosome.sizes}{Chromosome number in first column and corresponding size in second (BED format), can easily be created using samtools faidx command.}

\item{intervals}{Should each chromosome be split into intervals or proportions? Defaults to intervals.}

\item{size}{How many megabases for each interval or number of equal proportions to break chromosome into. Defaults to 10.}
}
\description{
This functiion takes a number or series of numbers (chromosome.sizes) in BED format and creates equal (intervals) of (size) megabases or (size) proportions. The chromosome.sizes can be created easily by using samtools faidx command.
}
\examples{
data <- data.frame("chromosome"=sequence(3), "size"=round(runif(3,5000000,50000000)))
create.genome.intervals(data, TRUE, 5)
}
\keyword{genome}
\keyword{intervals}

